reorganize marklogic package
