+++ src/main/java/uk/gov/legislation/converters/DocumentsFeedConverter.java
@@ -2,7 +2,7 @@ package uk.gov.legislation.converters;
 
 import uk.gov.legislation.api.responses.CommonMetadata;
 import uk.gov.legislation.api.responses.PageOfDocuments;
-import uk.gov.legislation.data.marklogic.SearchResults;
+import uk.gov.legislation.data.marklogic.search.SearchResults;
 import uk.gov.legislation.util.Cites;
 import uk.gov.legislation.util.Links;
