+++ src/main/java/uk/gov/legislation/endpoints/document/service/DocumentService.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 import uk.gov.legislation.api.responses.Document;
 import uk.gov.legislation.api.responses.DocumentMetadata;
 import uk.gov.legislation.converters.DocumentMetadataConverter;
-import uk.gov.legislation.data.marklogic.Legislation;
+import uk.gov.legislation.data.marklogic.legislation.Legislation;
 import uk.gov.legislation.endpoints.CustomHeaders;
 import uk.gov.legislation.exceptions.TransformationException;
 import uk.gov.legislation.transform.Akn2Html;
